tweaks on previous exercises

+++ modulo03/spring-III/exercicio01/src/main/java/br/com/meli/exercicio01/entity/Character.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Character {
     private String name;
-    private int height;
-    private int mass;
+    private Double height;
+    private Double mass;
     private String hairColor;
     private String skinColor;
     private String eyeColor;
